+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - `Innmind\BlackBox\Runner\Assert::debug()`
 - `Innmind\BlackBox\Set::disableShrinking()`
+- `Innmind\BlackBox\Runner\Proof::tagged()`
+- `Innmind\BlackBox\Runner\Proof::disableShrinking()`
+- `Innmind\BlackBox\Application::mapProof()`
 
 ### Changed